Handle clast_expr_name properly
CLooG 0.16.0 introduced clast_expr_name, which allows a name to appear anywhere
in an expression. Graphite did not yet support this. Without this patch Graphite
will fail, as soon as CLooG generates a clast_expr_name at a place which is not
a clast_expr_term->var.

Tobias Grosser <tobias@grosser.es>

	* graphite-clast-to-gimple.c (clast_name_to_index,
	clast_name_to_lb_ub, clast_name_to_gcc): Change types.
	(clast_to_gcc_expression): Add clast_expr_name as a new
	case. Do not assume a clast_expr_term points always to a
	clast_expr_name.
	(type_for_clast_term): Do not assume a clast_expr_term points always to
	a clast_expr_name.
	(type_for_clast_name): New.
	(type_for_clast_expr): Add clast_expr_name as a new case.