Add isl scheduling optimizer
The isl scheduling optimizer implements the scheduling algorithm first developed
in Pluto [1]. Pluto has shown significant speedups and is nowadays even
implemented in the IBM XL-C compiler. The implementation of this pass is a first
draft and was copied largely from Polly. We need still to adapt the code to the
gcc coding style and we need to tune the isl scheduler. At the moment we get
reasonable compile times (at most 2x-3x slowdown) and first speedups. We now
need to tune the compile time and start to investigate which optimizations and
heuristics need to be tuned in our reimplementation.

[1] http://pluto-compiler.sourceforge.net/

Tobias Grosser  <tobias@grosser.es>