Derive the size of the induction variables directly from the clast
With this we do not need the LST (loop structure tree) to derive the types of
the induction variables. Also, we do not derive them from the first statement in
the loop body (an arbitrary choice), but consider all statements. This is not
yet entirely correct, but better than we were before and as good as we can get
with the current implementation of CLooG.

Tobias Grosser  <tobias@grosser.es>

	* graphite-clast-to-gimple.c (compute_bounds_for_level): Removed.
	(clast_get_body_of_loop): Removed.
	(compute_bounds_for_loop): New.
	(graphite_create_new_loop): Use compute_bounds_for_loop instead of
	compute_bounds_for_level and clast_get_body_of_loop.
	(set_cloog_options): Ask CLooG to preserve the iteration domains in
	the clast.